% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGeoms.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline - this is the code to create the geom we'll use in our plot}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeline
}
\description{
GeomTimeline - this is the code to create the geom we'll use in our plot
}
\keyword{datasets}
