% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGeoms.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Theme for better timeline visualization in ggplot2}
\usage{
theme_timeline()
}
\description{
This is a simple theme that makes \code{\link{geom_timeline}}
look better.
}
\examples{
\dontrun{
getData() \%>\% eq_clean_data() \%>\%
filter(COUNTRY \%in\% c("USA"), YEAR > 2010) \%>\%
 ggplot(aes(x = DATE, y = COUNTRY, color = as.numeric(TOTAL_DEATHS), size = as.numeric(EQ_PRIMARY))) +
 geom_timeline() + theme_timeline() + labs(size = "Richter scale value", color = "# deaths")
}

}
