% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAndClean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data function takes raw NOAA data data frame and returns a data frame with the Date field added.}
\usage{
eq_clean_data(data)
}
\value{
A copy of the data data frame with the Date field added
}
\description{
eq_clean_data function takes raw NOAA data data frame and returns a data frame with the Date field added.
}
\examples{
{
\dontrun{
eq_clean_data(getData())
}}

}
