% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAndClean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data function takes raw NOAA signif data frame and returns a clean data frame.}
\usage{
eq_clean_data()
}
\value{
clean signif data frame
}
\description{
eq_clean_data function takes raw NOAA signif data frame and returns a clean data frame.
}
