% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAndClean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean function takes raw LOCATION NAME from the signif data frame and returns a clean field.}
\usage{
eq_location_clean()
}
\value{
LOCATION_NAME in title case and without the country name
}
\description{
eq_location_clean function takes raw LOCATION NAME from the signif data frame and returns a clean field.
}
