% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAndClean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean function takes raw LOCATION NAME from the data data frame and returns a clean field.}
\usage{
eq_location_clean(data)
}
\value{
modified data data frame with the LOCATION_NAME in title case and the country name removed from the LOCATION NAME field
}
\description{
eq_location_clean function takes raw LOCATION NAME from the data data frame and returns a clean field.
}
\examples{
{
\dontrun{
eq_location_clean(data)
}}

}
